linked tournament dashboard to the details page for presentation. Seperated my joint tables onto seperate js files and updated that in the index.js file
