+++ public/html/tournamentdashboard.html
@@ -176,7 +176,7 @@
                    <h2>Your Match against Scrub City did you?</h2>
 
                    <br>
-                   <button type="button"   href="/tournidashboard2.html" class="btn btn-dark">Win</button>
+                   <button type="button"  a href="/public/html/tournidashboard2.html" class="btn btn-dark">Win</button>
                    <br>
                    <br>
                    <button type="button" class="btn btn-dark">Lose</button>
