+++ views/playerhome.handlebars
@@ -196,7 +196,7 @@
                                 <td>Volleyball</td>
                                 <td>May 10 2018</td>
                                 <td>
-                                    <a href="details.html" class="btn btn-secondary">
+                                    <a href="/public/html/tournamentdashboard.html" class="btn btn-secondary">
                                         <i class="fas fa-angle-double-right"></i> Details
                                     </a>
                                 </td>
@@ -207,7 +207,7 @@
                                 <td>Bowling</td>
                                 <td>May 11 2018</td>
                                 <td>
-                                    <a href="details.html" class="btn btn-secondary">
+                                    <a href="/public/html/tournamentdashboard.html" class="btn btn-secondary">
                                         <i class="fas fa-angle-double-right"></i> Details
                                     </a>
                                 </td>
@@ -218,7 +218,7 @@
                                 <td>Softball</td>
                                 <td>May 13 2018</td>
                                 <td>
-                                    <a href="details.html" class="btn btn-secondary">
+                                    <a href="/public/html/tournamentdashboard.html" class="btn btn-secondary">
                                         <i class="fas fa-angle-double-right"></i> Details
                                     </a>
                                 </td>
@@ -229,7 +229,7 @@
                                 <td>Volleyball</td>
                                 <td>May 15 2018</td>
                                 <td>
-                                    <a href="details.html" class="btn btn-secondary">
+                                    <a href="/public/html/tournamentdashboard.html" class="btn btn-secondary">
                                         <i class="fas fa-angle-double-right"></i> Details
                                     </a>
                                 </td>
@@ -240,7 +240,7 @@
                                 <td>Basketball</td>
                                 <td>May 15 2018</td>
                                 <td>
-                                    <a href="details.html" class="btn btn-secondary">
+                                    <a href="/public/html/tournamentdashboard.html" class="btn btn-secondary">
                                         <i class="fas fa-angle-double-right"></i> Details
                                     </a>
                                 </td>
